requirements.txt
altair==5.5.0
asttokens==2.4.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloup==3.0.5
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.7
decorator==5.1.1
et_xmlfile==2.0.0
executing==2.1.0
fastjsonschema==2.21.0
fonttools==4.55.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
idna==3.10
imutils==0.5.4
ipykernel==6.29.5
ipython==8.28.0
ipython-genutils==0.2.0
isosurfaces==0.1.2
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lxml==5.3.0
manim==0.18.1
ManimPango==0.5.0
mapbox_earcut==1.0.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
moderngl==5.11.0
moderngl-window==2.4.6
multipledispatch==1.0.0
narwhals==1.14.3
nbformat==4.2.0
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
platformdirs==4.3.6
plotly==5.24.1
prompt_toolkit==3.0.48
protobuf==5.29.0
psutil==6.0.0
pure_eval==0.2.3
pyarrow==18.1.0
pycairo==1.26.1
pycparser==2.22
pydeck==0.9.1
pydub==0.25.1
pyglet==2.0.17
Pygments==2.18.0
pyparsing==3.2.0
pyrr==0.10.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
pytz==2024.2
pywin32==308
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.21.0
scipy==1.14.0
screeninfo==0.8.1
seaborn==0.13.2
selenium==4.26.1
six==1.16.0
skia-pathops==0.8.0.post1
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
srt==3.5.3
stack-data==0.6.3
streamlit==1.40.2
svgelements==1.9.6
tenacity==9.0.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==4.0.2
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
XlsxWriter==3.2.0
